+++ HelperStatic.variables.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Data;
 using System.IO;
 
 namespace GeoTagNinja;